@@ -1,20 +1,19 @@
 #!/usr/bin/env node
-'use strict';
 
-const {join} = require('path');
-const {cliBasics} = require('command-line-basics');
-const {makeBadgeFromJSONFile} = require('../src/makeBadge.js');
+import {fileURLToPath} from 'url';
+import {join, dirname} from 'path';
+import {cliBasics} from 'command-line-basics';
+import {makeBadgeFromJSONFile} from '../src/makeBadge.cjs';
 
-const optionDefinitions = cliBasics(
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const optionDefinitions = await cliBasics(
   join(__dirname, './optionDefinitions.js')
 );
 
 if (!optionDefinitions) { // cliBasics handled
-  process.exit();
+  process.exit(0);
 }
 
-// file, (options)
-(async () => {
 const {output} = await makeBadgeFromJSONFile(optionDefinitions);
 console.log('Saved to ' + output);
-})();
