@@ -12,6 +12,7 @@ module.exports = {
         }
     }],
     parserOptions: {
-        ecmaVersion: 2017
+        sourceType: 'module',
+        ecmaVersion: 2022
     }
 };
